simulateur/requirements.txt
anyio
asgiref
attrs
bleach
channels
channels-redis
Django
django-json-widget
djangorestframework
httpx
idna
Jinja2
MarkupSafe
matplotlib
matplotlib-inline
numpy
packaging
pandas
Pygments
pytz
PyYAML
redis
requests
sqlparse
traitlets
urllib3
